requirements.txt
fastapi
uvicorn
Pillow
ultralytics
python-multipart